# frontend/tooling/codegen/swagger_original.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
...
        "/api/admin/shop": {
            "get": {
                "tags": [
                    "shop-admin-controller"
                ],
                "summary": "listAll",
                "operationId": "listAllUsingGET",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ShopsAdminDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/api/admin/shop/{id}": {
            "get": {
                "tags": [
                    "shop-admin-controller"
                ],
                "summary": "getShopSettings",
                "operationId": "getShopSettingsUsingGET",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ShopAdminDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "put": {
                "tags": [
                    "shop-admin-controller"
                ],
                "summary": "update",
                "operationId": "updateUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "request",
                        "description": "request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateShopDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ShopAdminDto"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": [
                    "shop-admin-controller"
                ],
                "summary": "delete",
                "operationId": "deleteUsingDELETE",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "deprecated": false
            }
        },
        "/api/admin/shop/{id}/approve": {
            "put": {
                "tags": [
                    "shop-admin-controller"
                ],
                "summary": "changeApprove",
                "operationId": "changeApproveUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "approved",
                        "in": "query",
                        "description": "approved",
                        "required": true,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
        "/api/shop/me": {
            "get": {
                "tags": [
                    "shop-owner-controller"
                ],
                "summary": "getShopSettings",
                "operationId": "getShopSettingsUsingGET_1",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ShopOwnerDetailDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
        "/api/shop/{id}": {
            "get": {
                "tags": [
                    "shop-controller"
                ],
                "summary": "getShopDetails",
                "operationId": "getShopDetailsUsingGET",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ShopDetailDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
        "BreaksDto": {
            "type": "object",
            "properties": {
                "slotIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            },
            "title": "BreaksDto"
        },
...
        "DayDto": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            },
            "title": "DayDto"
        },
...
        "ShopAdminDto": {
            "type": "object",
            "properties": {
                "addressSupplement": {
                    "type": "string"
                },
                "approved": {
                    "type": "boolean"
                },
                "autoColorEnabled": {
                    "type": "boolean"
                },
                "city": {
                    "type": "string"
                },
                "contacts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "details": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerName": {
                    "type": "string"
                },
                "shareLink": {
                    "type": "string"
                },
                "slots": {
                    "$ref": "#/definitions/SlotConfigDto"
                },
                "socialLinks": {
                    "$ref": "#/definitions/SocialLinksDto"
                },
                "street": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            },
            "title": "ShopAdminDto"
        },
        "ShopDetailDto": {
            "type": "object",
            "properties": {
                "addressSupplement": {
                    "type": "string"
                },
                "autoColorEnabled": {
                    "type": "boolean"
                },
                "city": {
                    "type": "string"
                },
                "contactTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "details": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerName": {
                    "type": "string"
                },
                "shareLink": {
                    "type": "string"
                },
                "shopColor": {
                    "type": "string"
                },
                "socialLinks": {
                    "$ref": "#/definitions/SocialLinksDto"
                },
                "street": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            },
            "title": "ShopDetailDto"
        },
...
        "ShopOwnerDetailDto": {
            "type": "object",
            "properties": {
                "addressSupplement": {
                    "type": "string"
                },
                "autoColorEnabled": {
                    "type": "boolean"
                },
                "breaks": {
                    "$ref": "#/definitions/BreaksDto"
                },
                "city": {
                    "type": "string"
                },
                "contacts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerName": {
                    "type": "string"
                },
                "shareLink": {
                    "type": "string"
                },
                "slots": {
                    "$ref": "#/definitions/SlotConfigDto"
                },
                "socialLinks": {
                    "$ref": "#/definitions/SocialLinksDto"
                },
                "street": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            },
            "title": "ShopOwnerDetailDto"
        },
        "ShopsAdminDto": {
            "type": "object",
            "properties": {
                "shops": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ShopAdminDto"
                    }
                }
            },
            "title": "ShopsAdminDto"
        },
        "SlotConfigDto": {
            "type": "object",
            "properties": {
                "delayForFirstSlot": {
                    "type": "integer",
                    "format": "int32"
                },
                "friday": {
                    "$ref": "#/definitions/DayDto"
                },
                "monday": {
                    "$ref": "#/definitions/DayDto"
                },
                "saturday": {
                    "$ref": "#/definitions/DayDto"
                },
                "sunday": {
                    "$ref": "#/definitions/DayDto"
                },
                "thursday": {
                    "$ref": "#/definitions/DayDto"
                },
                "timeBetweenSlots": {
                    "type": "integer",
                    "format": "int32"
                },
                "timePerSlot": {
                    "type": "integer",
                    "format": "int32"
                },
                "tuesday": {
                    "$ref": "#/definitions/DayDto"
                },
                "wednesday": {
                    "$ref": "#/definitions/DayDto"
                }
            },
            "title": "SlotConfigDto"
        },
...
        "SocialLinksDto": {
            "type": "object",
            "properties": {
                "facebook": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "twitter": {
                    "type": "string"
                }
            },
            "title": "SocialLinksDto"
        },
...
        "UpdateShopDto": {
            "type": "object",
            "properties": {
                "addressSupplement": {
                    "type": "string"
                },
                "autoColorEnabled": {
                    "type": "boolean"
                },
                "breaks": {
                    "$ref": "#/definitions/BreaksDto"
                },
                "city": {
                    "type": "string"
                },
                "contacts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "details": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerName": {
                    "type": "string"
                },
                "slots": {
                    "$ref": "#/definitions/SlotConfigDto"
                },
                "socialLinks": {
                    "$ref": "#/definitions/SocialLinksDto"
                },
                "street": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            },
            "title": "UpdateShopDto"
        },
...
    }
}